Creación de Formularios
Se crean Formularios de tipo imagen, Eps, rh y tipo documento

+++ src/HCRadiologia/Persona.java
@@ -167,7 +167,7 @@ public class Persona {
 		intRH = newVal;
 	}
 
-	public byte getByteGenero(){
+	public int getByteGenero(){
             
 		return byteGenero;
 	}
@@ -353,6 +353,8 @@ public class Persona {
 	   }
             
 	}
+        
+        
 //==========================================================================================================================
 // Main para hacer pruebas
